requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.51.0
python-dotenv==1.0.1
httpx==0.27.2
python-telegram-bot==21.6
pymupdf==1.24.11